requirements.txt
# requirements.txt - Version optimisée pour votre projet

# Interface web
streamlit==1.28.0

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0

# Traitement d'images
pillow==10.0.0

# Data Science
numpy==1.24.0
pandas==2.0.0
scipy==1.11.0

# Visualisation
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0

# Base de données (inclus avec Python)
# sqlite3